refactor(gui): 优化分区和区域下拉框的默认值设置
- 将 selected_part_index 和 selected_area_index 的类型从 size_t 改为 int
- 在设置默认值时使用 static_cast<int> 进行类型转换
- 提高代码的可读性和类型安全性

+++ src/plugin-main.cpp
@@ -380,11 +380,11 @@ public slots:
         }
 
         // 填充分区下拉框并设置默认值
-        size_t selected_part_index = 0;
+        int selected_part_index = 0;
         for (size_t i = 0; i < parts.size(); i++) {
             partCombo->addItem(parts[i].name, parts[i].id);
             if (parts[i].id == config.part_id) {
-                selected_part_index = i;
+                selected_part_index = static_cast<int>(i);
             }
         }
         partCombo->setCurrentIndex(selected_part_index);
@@ -402,7 +402,7 @@ public slots:
                     QString name = QString::fromUtf8(obs_data_get_string(area, "name"));
                     areaCombo->addItem(name, id);
                     if (id == config.area_id) {
-                        selected_area_index = i;
+                        selected_area_index = static_cast<int>(i);
                     }
                     obs_data_release(area);
                 }
